moved spinnaker.cfg back to its traditional place
