+++ spinn_front_end_common/interface/spinnaker.cfg.template
