+++ unittests/test_cfg_checker.py
@@ -16,7 +16,7 @@
 import os
 import unittest
 from spinn_utilities.config_holder import check_python_file
-from spinn_front_end_common.config_setup import reset_configs
+from spinn_front_end_common.interface.config_setup import reset_configs
 
 
 class TestCfgChecker(unittest.TestCase):
